requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1

pillow==11.1.0


drf-spectacular==0.28.0

sentry-sdk==2.20.0

django-environ==0.12.0
python-dotenv==1.0.1
django-query-inspector==1.3.0

django-jazzmin==3.0.1

django-cors-headers==4.7.0
